'continue', 'next', & 'prev' all work flawlessly now (see notes).
Being able to suspend tracer execution by running the tracer on a separate
thread helped a lot in making this possible. No doubt, we can still refactor,
for one i don't like how the 'app' object is passed around (sort of hacky) but
the main goal for this commit was to make navigation commands stable.

The tracer is executed with the #continue method, it immediately suspends
itself, waiting to be told to run again. This makes it very easy to implement
the nav commands.