Handle set_trace_func crashes gracefully.
Thanks to @envygeeks for spotting this one, he had written code which would
infinitely recurse and cause irwebmachine to lock up because set_trace_func
would be called infinitely as well. We can't handle this in any other way but
to crash & stop the tracer. The stack up to the point of the infinite recursion
is kept.

Exceptions that do not cause set_trace_func to infinitely recurse are not
vulernable to this one.